+++ libs/web/calendar/dialogs/google-event-dialog/feature/src/lib/google-event-dialog.module.ts
@@ -2,13 +2,19 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { ReactiveComponentModule } from '@ngrx/component';
-import { TuiButtonModule, TuiErrorModule } from '@taiga-ui/core';
+import { TuiMapperPipeModule } from '@taiga-ui/cdk';
+import {
+  TuiButtonModule,
+  TuiDataListModule,
+  TuiErrorModule,
+} from '@taiga-ui/core';
 import {
   TuiCheckboxLabeledModule,
   TuiFieldErrorPipeModule,
   TuiInputDateModule,
   TuiInputModule,
   TuiInputTimeModule,
+  TuiSelectModule,
   TuiTextAreaModule,
 } from '@taiga-ui/kit';
 import { GoogleEventDialogComponent } from './google-event-dialog.component';
@@ -17,11 +23,14 @@ const NGRX = [ReactiveComponentModule];
 const TAIGA_UI = [
   TuiButtonModule,
   TuiCheckboxLabeledModule,
+  TuiDataListModule,
   TuiErrorModule,
   TuiFieldErrorPipeModule,
   TuiInputDateModule,
   TuiInputModule,
   TuiInputTimeModule,
+  TuiMapperPipeModule,
+  TuiSelectModule,
   TuiTextAreaModule,
 ];